Consolidate dev_shm_nodev/noexec/nosuid into dev_shm, fix without fabricating fstab
dev_shm_nodev/noexec/nosuid duplicated what the existing dev_shm class
already does by default in every OS bundle, so fold them into dev_shm
and delete the separate classes.

Also fixes how dev_shm itself applies these options. It previously used
file_line to force /etc/fstab into a single hardcoded line, fabricating
an entry when none existed and clobbering any options an admin had
already set (size, mode, etc). /dev/shm is a systemd "API filesystem"
(like /proc, /sys, /run) - systemd refuses to let a .mount unit be
created for it (verified: "Cannot create mount unit for API file
system /dev/shm. Refusing."), and systemd's own mechanism for adjusting
these mounts (systemd-remount-fs.service) reads from /etc/fstab, so
there's no "systemd instead of fstab" path available here. dev_shm now
only acts when a real, pre-existing fstab entry is found, and edits it
incrementally via the same set_mount_options define the other mount
option rules use - non-destructive to any options already present,
and a documented no-op on hosts without a real entry instead of
fabricating one.
