+++ data/cis/cis_AlmaLinux_8_params.yaml
@@ -38,9 +38,6 @@ cis_security_hardening::rules::var_log_audit_noexec::enforce: true
 cis_security_hardening::rules::var_log_audit_nosuid::enforce: true
 cis_security_hardening::rules::dev_shm::enforce: true
 cis_security_hardening::rules::dev_shm::size: 0
-cis_security_hardening::rules::dev_shm_noexec::enforce: true
-cis_security_hardening::rules::dev_shm_nodev::enforce: true
-cis_security_hardening::rules::dev_shm_nosuid::enforce: true
 cis_security_hardening::rules::var_tmp_noexec::enforce: true
 cis_security_hardening::rules::var_tmp_nodev::enforce: true
 cis_security_hardening::rules::var_tmp_nosuid::enforce: true
