+++ data/cis/cis_CentOS_7_params.yaml
@@ -35,9 +35,6 @@ cis_security_hardening::rules::tmp_nodev::enforce: true
 cis_security_hardening::rules::tmp_nosuid::enforce: true
 cis_security_hardening::rules::dev_shm::enforce: true
 cis_security_hardening::rules::dev_shm::size: 0
-cis_security_hardening::rules::dev_shm_noexec::enforce: true
-cis_security_hardening::rules::dev_shm_nodev::enforce: true
-cis_security_hardening::rules::dev_shm_nosuid::enforce: true
 cis_security_hardening::rules::var_tmp_noexec::enforce: true
 cis_security_hardening::rules::var_tmp_nodev::enforce: true
 cis_security_hardening::rules::var_tmp_nosuid::enforce: true
