+++ data/cis/cis_CentOS_8_rules.yaml
@@ -27,9 +27,6 @@ cis_security_hardening::benchmark::centos::v8:
         - home_usrquota
         - home_grpquota
         - dev_shm
-        - dev_shm_noexec
-        - dev_shm_nodev
-        - dev_shm_nosuid
         - disable_automount
         - disable_usb_storage
     software_updates:
