+++ data/cis/cis_Debian_10_rules.yaml
@@ -15,9 +15,6 @@ cis_security_hardening::benchmark::debian::v10:
         - tmp_nosuid
         - tmp_noexec
         - dev_shm
-        - dev_shm_nodev
-        - dev_shm_nosuid
-        - dev_shm_noexec
         - var_tmp_nodev
         - var_tmp_nosuid
         - var_tmp_noexec
