+++ data/cis/cis_Debian_12_params.yaml
@@ -42,9 +42,6 @@ cis_security_hardening::rules::home_nodev::enforce: true
 cis_security_hardening::rules::home_nosuid::enforce: true
 cis_security_hardening::rules::dev_shm::enforce: true
 cis_security_hardening::rules::dev_shm::size: 0
-cis_security_hardening::rules::dev_shm_noexec::enforce: true
-cis_security_hardening::rules::dev_shm_nodev::enforce: true
-cis_security_hardening::rules::dev_shm_nosuid::enforce: true
 cis_security_hardening::rules::sticky_world_writeable_files::enforce: true
 cis_security_hardening::rules::disable_automount::enforce: true
 cis_security_hardening::rules::disable_usb_storage::enforce: true
