+++ data/cis/cis_RedHat_8_rules.yaml
@@ -27,9 +27,6 @@ cis_security_hardening::benchmark::redhat::v8:
         - home_usrquota
         - home_grpquota
         - dev_shm
-        - dev_shm_noexec
-        - dev_shm_nodev
-        - dev_shm_nosuid
         - sticky_world_writeable_files
         - disable_automount
         - disable_usb_storage
