+++ data/cis/cis_Ubuntu_22.04_params.yaml
@@ -42,9 +42,6 @@ cis_security_hardening::rules::home_nodev::enforce: true
 cis_security_hardening::rules::home_nosuid::enforce: true
 cis_security_hardening::rules::dev_shm::enforce: true
 cis_security_hardening::rules::dev_shm::size: 0
-cis_security_hardening::rules::dev_shm_noexec::enforce: true
-cis_security_hardening::rules::dev_shm_nodev::enforce: true
-cis_security_hardening::rules::dev_shm_nosuid::enforce: true
 
 cis_security_hardening::rules::apt_unused::enforce: true
 cis_security_hardening::rules::apt_unused::files:
