+++ data/cis/cis_Ubuntu_24.04_rules.yaml
@@ -30,9 +30,6 @@ cis_security_hardening::benchmark::ubuntu::v24:
         - home_nodev
         - home_nosuid
         - dev_shm
-        - dev_shm_nodev
-        - dev_shm_nosuid
-        - dev_shm_noexec
         - sticky_world_writeable_files
         - disable_automount
     filesystem_integrety:
