+++ manifests/rules/dev_shm.pp
@@ -15,7 +15,7 @@
 #    Enforce the rule
 #
 # @param size
-#    Size in GB
+#    Size in GB. 0 leaves the existing size untouched.
 #
 # @example
 #   class { 'cis_security_hardening::rules::dev_shm':
@@ -23,24 +23,45 @@
 #   }
 #
 # @api private
+#
+# @note
+#    /dev/shm is one of systemd's "API filesystems" (like /proc, /sys, /run) -
+#    it is mounted directly by PID1's early startup code, and systemd
+#    explicitly refuses to create a .mount unit for it. The only mechanism
+#    systemd itself provides for customizing its options
+#    (systemd-remount-fs.service) reads from /etc/fstab, so this rule only
+#    applies when /dev/shm already has a real, pre-existing fstab entry - it
+#    edits that entry incrementally (rather than replacing it outright) so
+#    any other options an administrator has already set are preserved. Hosts
+#    without a real entry are left untouched rather than having one
+#    fabricated for them.
 class cis_security_hardening::rules::dev_shm (
   Boolean $enforce = false,
   Integer $size    = 0,
 ) {
-  if $enforce {
-    if $size == 0 {
-      $options = 'defaults,nodev,nosuid,noexec,seclabel'
-    } else {
-      $options = "defaults,size=${size}G,nodev,nosuid,noexec,seclabel"
+  if ($enforce) and cis_security_hardening::hash_key($facts['mountpoints'], '/dev/shm') and
+  cis_security_hardening::hash_key($facts['cis_security_hardening'], 'fstab_mountpoints') and
+  cis_security_hardening::hash_key($facts['cis_security_hardening']['fstab_mountpoints'], '/dev/shm') {
+    cis_security_hardening::set_mount_options { '/dev/shm-nodev':
+      mountpoint   => '/dev/shm',
+      mountoptions => 'nodev',
     }
 
-    $line = "tmpfs   /dev/shm        tmpfs   ${options}   0 0"
-    file_line { 'add /dev/shm to fstab':
-      ensure             => present,
-      path               => '/etc/fstab',
-      match              => "^tmpfs\\s* /dev/shm",
-      line               => $line,
-      append_on_no_match => true,
+    cis_security_hardening::set_mount_options { '/dev/shm-noexec':
+      mountpoint   => '/dev/shm',
+      mountoptions => 'noexec',
+    }
+
+    cis_security_hardening::set_mount_options { '/dev/shm-nosuid':
+      mountpoint   => '/dev/shm',
+      mountoptions => 'nosuid',
+    }
+
+    if $size > 0 {
+      cis_security_hardening::set_mount_options { '/dev/shm-size':
+        mountpoint   => '/dev/shm',
+        mountoptions => "size=${size}G",
+      }
     }
   }
 }
